fix rest of tests, preparing to add more coverage later
